rewrite copy-to-clipboard function as a new component
